Fixed CODEOWNERS for USM
