+++ pkg/network/ebpf/kprobe_types.go
@@ -11,7 +11,6 @@ package ebpf
 /*
 #include "./c/tracer.h"
 #include "./c/tcp_states.h"
-#include "./c/tags-types.h"
 #include "./c/prebuilt/offset-guess.h"
 */
 import "C"
@@ -47,17 +46,3 @@ const (
 )
 
 const BatchSize = C.CONN_CLOSED_BATCH_SIZE
-
-type ConnTag = uint64
-
-const (
-	GnuTLS  ConnTag = C.LIBGNUTLS
-	OpenSSL ConnTag = C.LIBSSL
-)
-
-var (
-	StaticTags = map[ConnTag]string{
-		GnuTLS:  "tls.library:gnutls",
-		OpenSSL: "tls.library:openssl",
-	}
-)
