+++ pkg/network/http/http_types_linux.go
@@ -64,3 +64,17 @@ const (
 
 	libPathMaxSize = 0x78
 )
+
+type ConnTag = uint64
+
+const (
+	GnuTLS  ConnTag = 0x1
+	OpenSSL ConnTag = 0x2
+)
+
+var (
+	StaticTags = map[ConnTag]string{
+		GnuTLS:  "tls.library:gnutls",
+		OpenSSL: "tls.library:openssl",
+	}
+)
